+++ hav-imobiliaria/src/main/java/com/hav/hav_imobiliaria/service/EditorService.java
@@ -43,7 +43,6 @@ public class EditorService {
         // que mudar todas as dtos e front end para adicionar o user_details
 
 
-
         Editor savedEditor = repository.save(editor);
 
         if (image != null) {
@@ -64,11 +63,8 @@ public class EditorService {
 
         modelMapper.map(editorPutDTO, editor);
 
-        if (deletedImageId != null) {
+        if (deletedImageId != null && newImage != null) {
             imageService.deleteUserImage(deletedImageId);
-        }
-
-        if (newImage != null) {
             imageService.uploadUserImage(id, newImage);
         }
 
@@ -112,7 +108,12 @@ public class EditorService {
         Editor editor = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Realtor not found"));
 
-        // Converte a entidade Realtor para o DTO
-        return modelMapper.map(editor, EditorPutRequestDTO.class);
+        EditorPutRequestDTO editorPutRequestDTO = modelMapper.map(editor, EditorPutRequestDTO.class);
+
+        if (editor.getImageUser() != null) {
+            editorPutRequestDTO.setImageId(editor.getImageUser().getId());
+        }
+
+        return editorPutRequestDTO;
     }
 }
