+++ hav-imobiliaria/src/main/java/com/hav/hav_imobiliaria/service/RealtorService.java
@@ -7,6 +7,7 @@ import com.hav.hav_imobiliaria.model.entity.Users.Adm;
 import com.hav.hav_imobiliaria.model.entity.Users.Realtor;
 import com.hav.hav_imobiliaria.model.entity.Users.User;
 import com.hav.hav_imobiliaria.repository.RealtorRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
@@ -61,11 +62,8 @@ public class RealtorService {
 
         modelMapper.map(realtorPutDTO, realtor);
 
-        if (deletedImageId != null) {
+        if (deletedImageId != null && newImage != null) {
             imageService.deleteUserImage(deletedImageId);
-        }
-
-        if (newImage != null) {
             imageService.uploadUserImage(id, newImage);
         }
 
@@ -162,10 +160,16 @@ public class RealtorService {
     }
 
     public RealtorPutRequestDTO findRealtorById(Integer id) {
-        Realtor realtor = repository.findById(id).get();
+        Realtor realtor = repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Realtor not found"));
 
-        // Converte a entidade Realtor para o DTO
-        return modelMapper.map(realtor, RealtorPutRequestDTO.class);
+        RealtorPutRequestDTO realtorDTO = modelMapper.map(realtor, RealtorPutRequestDTO.class);
+
+        if (realtor.getImageUser() != null) {
+            realtorDTO.setImageId(realtor.getImageUser().getId());
+        }
+
+        return realtorDTO;
     }
 
 }
